Fixed width integers

+++ src/cue/show.cpp
@@ -12,7 +12,7 @@ namespace PixelMaestro {
 	 * @param events Array of Events to queue.
 	 * @param num_events The number of Events in the queue.
 	 */
-	Show::Show(CueController* controller, Event* events, unsigned short num_events) {
+	Show::Show(CueController* controller, Event* events, uint16_t num_events) {
 		cue_controller_ = controller;
 		set_events(events, num_events);
 	}
@@ -41,7 +41,7 @@ namespace PixelMaestro {
 		@param num_events The number of Events in the queue.
 		@param preserve_current_index If false, reset the current Event index to 0. Defaults to true.
 	*/
-	void Show::set_events(Event* events, unsigned short num_events, bool preserve_current_index) {
+	void Show::set_events(Event* events, uint16_t num_events, bool preserve_current_index) {
 		if (!preserve_current_index) {
 			current_index_ = 0;
 		}
@@ -72,7 +72,7 @@ namespace PixelMaestro {
 
 		@param current_time Program runtime.
 	*/
-	void Show::update(const unsigned long& current_time) {		
+	void Show::update(const uint32_t& current_time) {		
 		// Only run if we're looping, or if we haven't reached the end of the Event list yet.
 		if (loop_ || (!loop_ && current_index_ != num_events_)) {
 			check_next_event(current_time);
@@ -84,14 +84,14 @@ namespace PixelMaestro {
 	/**
 	 * Checks the next Event's start time, then runs it if it's ready.
 	 */
-	void Show::check_next_event(const unsigned long& current_time) {
+	void Show::check_next_event(const uint32_t& current_time) {
 		/*
 			Based on the timing method used, determine whether to run the Event.
 			If ABSOLUTE, compare the current time to the next Event's start time.
 			If RELATIVE, compare the time since the last Event to the next Event's start time.
 			After running the Event, update the last run time and current Event index.
 		*/
-		unsigned long event_time = events_[current_index_].get_time();
+		uint32_t event_time = events_[current_index_].get_time();
 		if ((timing_ == TimingMode::Absolute && (current_time >= event_time)) ||
 			(timing_ == TimingMode::Relative && ((current_time - last_time_) >= event_time))) {
 			events_[current_index_].run(cue_controller_);
