+++ src/cue/show.h
@@ -22,37 +22,37 @@ namespace PixelMaestro {
 				Relative
 			};
 
-			Show(CueController* controller, Event *events, unsigned short num_events);
+			Show(CueController* controller, Event *events, uint16_t num_events);
 			bool get_looping();
 			TimingMode get_timing();
-			void set_events(Event* events, unsigned short num_events, bool preserve_current_index = true);
+			void set_events(Event* events, uint16_t num_events, bool preserve_current_index = true);
 			void set_timing(TimingMode timing);
 			void set_looping(bool loop);
-			void update(const unsigned long& current_time);
+			void update(const uint32_t& current_time);
 
 		private:
 			/// The controller for running Cues.
 			CueController* cue_controller_ = nullptr;
 
 			/// The index of the current Event.
-			unsigned short current_index_ = 0;
+			uint16_t current_index_ = 0;
 
 			/// Events used in the Show.
 			Event *events_;
 
 			/// The time that the last Event ran.
-			unsigned long last_time_ = 0;
+			uint32_t last_time_ = 0;
 
 			/// Whether to loop over the Event.
 			bool loop_ = false;
 
 			/// The number of Events in the Show.
-			unsigned short num_events_;
+			uint16_t num_events_;
 
 			/// Method for measuring a Event's start time. Defaults to Absolute.
 			TimingMode timing_ = TimingMode::Absolute;
 
-			void check_next_event(const unsigned long& current_time);
+			void check_next_event(const uint32_t& current_time);
 			void update_event_index();
 	};
 }
